+++ packages/tiny-canvas/src/index.js
@@ -6,4 +6,5 @@ export { default as Frame } from './Frame';
 export { default as Note } from './Note';
 export { default as Mark } from './Mark';
 export { default as Link } from './Link';
+export { default as Image } from './Image';
 export { useResetCanvas } from './useResetCanvas';